Week 3 | Laboratory Activity 3
The code represents basic vehicle information, categorized into three types: Car, Electric Car, and Motorcycle. Each type of vehicle is defined as a separate class with unique characteristics. The Car class includes a property $ndoors to specify the number of doors for a given car model. The ElectricCar class extends the Car class and adds a $batteryVolume property to indicate the car's battery capacity, while also inheriting the $ndoors property (now protected to allow access in subclasses). Finally, the Motorcycle class introduces a $variety property to specify the type of motorcycle model.